\name{VertexSort-package}

\alias{VertexSort-package}

\docType{package}

\title{The Vertex Sort Package}

\description{
  The present package is used to elucidate the hierarchical structure of a
  directed network using the Vertex Sort algorithm (Jothi et al., 2009).
  This package includes also functions for randomizing a graph in five
  different ways.
}

\details{
  \tabular{ll}{
    Package: \tab VertexSort\cr
    Type: \tab Package\cr
    Version: \tab 0.1-1\cr
    Date: \tab 2017-03-07\cr
    License: \tab GPL (>= 2)\cr
  }
}

\author{
  Diala Abd-Rabbo
  Maintainer: Diala Abd-Rabbo <diala.abd.rabbo@gmail.com>
}

\references{
  Jothi, R., Balaj, S., Wuster, A. et al. 2009 \emph{Molecular system
    biology} \bold{5}, --294-309.
}

\keyword{ package }
